+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+<<<<<<< HEAD
 **NetTool Beta 0.9.1**
 
 **Release 2026-06-22**
@@ -54,3 +55,61 @@ Initial Beta Release
 
 \-More features planned
 
+=======
+**NetTool Beta 0.9.1**
+
+**Release 2026-06-22**
+
+
+
+Initial Beta Release
+
+
+
+**Features:**
+
+\- Network summary command
+
+\-DNS lookups (A, AAAA, MX, CNAME, TXT, NS)
+
+\-Reverse DNS Lookups
+
+\-IP inspection (RDAP + ASN)
+
+\-ASN-enhanced traceroute
+
+\-Subnet calculator
+
+\-Integrated help system
+
+\-About command
+
+
+
+**0.9.0.1 fixes:**
+
+\-Added new tags for trace -o
+
+&#x09;examples:
+
+&#x09;	192.0.0.0/24 - Special-Purpose Address
+
+&#x09;	100.64.0.0/10 - Carrier-Grade NAT
+
+
+
+**0.9.1 fixes**
+
+\-Fixed Subnet calculator to display right output for /31 and /32 subnet blocks
+
+\-Added more sources for ASN lookups
+
+
+
+**Limitations**
+
+\-Windows only (Linux support coming soon)
+
+\-More features planned
+
+>>>>>>> 86b9eb6969b796b62ee614f2ae224f95c95f22ff